+++ RevoProfit.Core/Stock/Models/StockOwned.cs
@@ -3,10 +3,10 @@
 public class StockOwned
 {
     public required string Ticker { get; init; }
-    public double Quantity { get; set; }
-    public double AveragePrice { get; set; }
-    public double ValueInserted { get; set; }
-    public double TotalDividend { get; set; }
+    public decimal Quantity { get; set; }
+    public decimal AveragePrice { get; set; }
+    public decimal ValueInserted { get; set; }
+    public decimal TotalDividend { get; set; }
     public override string ToString()
     {
         return $"{Ticker}, Quantity: {Math.Round(Quantity, 14, MidpointRounding.ToEven)}, AveragePrice: ${Math.Round(AveragePrice, 2, MidpointRounding.ToEven)}, ValueInserted: ${Math.Round(ValueInserted, 2, MidpointRounding.ToEven)}, TotalDividend: ${Math.Round(TotalDividend, 2, MidpointRounding.ToEven)}";
